requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

numpy<2.0
pandas
scipy
matplotlib
pyyaml
future
scikit-image
opencv-python
Pillow
astropy
reproject
astroalign
torch>=2.0.0
torchvision
torchaudio
tensorboard 
einops
timm
nvitop 
tqdm
lmdb
addict
yapf
torch>=1.9.0
torchvision
tensorboard
tqdm
numpy
Pillow
timm
